third_party/WebKit/LayoutTests/imported/web-platform-tests/dom/nodes/Document-contentType/contentType/createDocument-expected.txt

This is a testharness.js-based test.
FAIL document.implementation.createDocument: document.contentType === 'application/xml' assert_equals: expected "application/xml" but got "application/xhtml+xml"
Harness: the test ran to completion.
